refactor job actions and kwargs
